server/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
WebTest
factory-boy>=2.2.1
fake-factory

# Management script
Flask-Script
konch
